pyproject.toml
[project]
name = "mona"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.11.3",
    "duckduckgo-search>=8.1.1",
    "llama-cpp-python>=0.3.16",
    "pypdf>=6.8.0",
    "pypdf2>=3.0.1",
    "pyttsx3>=2.99",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
]
